• split editor to different files. • Remove unneeded gets
