+++ src/cards/horizontal-buttons-stack/editor.js
@@ -1,0 +1,105 @@
+import { html } from "lit";
+
+export function renderHorButtonStackEditor(editor){
+    if (!editor.buttonAdded) {
+        editor.buttonAdded = true;
+        editor.buttonIndex = 0;
+
+        while (editor._config[(editor.buttonIndex + 1) + '_link']) {
+            editor.buttonIndex++;
+        }
+    }
+
+    function addButton() {
+        editor.buttonIndex++;
+        editor.requestUpdate();
+    }
+
+    return html`
+        <div class="card-config">
+            ${editor.makeDropdown("Card type", "card_type", editor.cardTypeList)}
+            <div id="buttons-container">
+                ${editor.makeButton()}
+            </div>
+            <button class="icon-button" @click="${addButton}">
+                <ha-icon icon="mdi:plus"></ha-icon>
+                New button
+            </button>
+            <ha-formfield .label="Auto order">
+                <ha-switch
+                    aria-label="Toggle auto order"
+                    .checked=${editor._config?.auto_order || false}
+                    .configValue="${"auto_order"}"
+                    @change=${editor._valueChanged}
+                ></ha-switch>
+                <div class="mdc-form-field">
+                    <label class="mdc-label">Optional - Auto order (Presence/occupancy sensors needed)</label> 
+                </div>
+            </ha-formfield>
+            <ha-expansion-panel outlined>
+                <h4 slot="header">
+                  <ha-icon icon="mdi:palette"></ha-icon>
+                  Styling options
+                </h4>
+                <div class="content">  
+                    <ha-expansion-panel outlined>
+                        <h4 slot="header">
+                          <ha-icon icon="mdi:palette"></ha-icon>
+                          Horizontal buttons stack styling
+                        </h4>
+                        <div class="content"> 
+                            <ha-textfield
+                                label="Optional - Margin (fix centering on some themes) (e.g. 13px)"
+                                .value="${editor._config?.margin || '7px'}"
+                                .configValue="${"margin"}"
+                                @input="${editor._valueChanged}"
+                            ></ha-textfield>
+                            <ha-textfield
+                                label="Optional - Width on desktop (100% by default on mobile)"
+                                .value="${editor._config?.width_desktop || '540px'}"
+                                .configValue="${"width_desktop"}"
+                                @input="${editor._valueChanged}"
+                            ></ha-textfield>
+                            <ha-formfield .label="Optional - Rise animation (Displays an animation once the page has loaded)">
+                                <ha-switch
+                                    aria-label="Optional - Rise animation (Displays an animation once the page has loaded)"
+                                    .checked=${editor._config?.rise_animation !== undefined ? editor._config?.rise_animation : true}
+                                    .configValue="${"rise_animation"}"
+                                    @change=${editor._valueChanged}
+                                ></ha-switch>
+                                <div class="mdc-form-field">
+                                    <label class="mdc-label">Optional - Rise animation (Displays an animation once the page has loaded)</label> 
+                                </div>
+                            </ha-formfield>
+                            <ha-formfield .label="Optional - Highlight current hash / view">
+                                <ha-switch
+                                    aria-label="Optional - Highlight current hash / view"
+                                    .checked=${editor._config?.highlight_current_view || false}
+                                    .configValue="${"highlight_current_view"}"
+                                    @change=${editor._valueChanged}
+                                ></ha-switch>
+                                <div class="mdc-form-field">
+                                    <label class="mdc-label">Optional - Highlight current hash / view</label> 
+                                </div>
+                            </ha-formfield>
+                            <ha-formfield .label="Optional - Hide gradient">
+                                <ha-switch
+                                    aria-label="Optional - Hide gradient"
+                                    .checked=${editor._config.hide_gradient || false}
+                                    .configValue="${"hide_gradient"}"
+                                    @change=${editor._valueChanged}
+                                ></ha-switch>
+                                <div class="mdc-form-field">
+                                    <label class="mdc-label">Optional - Hide gradient</label> 
+                                </div>
+                            </ha-formfield>
+                        </div>
+                    </ha-expansion-panel>
+                    ${editor.makeStyleEditor()}
+                </div>
+            </ha-expansion-panel>
+            <ha-alert alert-type="info">This card is the companion to the pop-up card, allowing you to open the corresponding pop-ups. It also allows you to open any page of your dashboard. In addition, you can add your motion sensors so that the order of the buttons adapts according to the room you just entered. This card is scrollable, remains visible and acts as a footer.</ha-alert>
+            ${editor.makeVersion()}
+        </div>
+    `;
+}
