+++ src/editor/bubble-card-editor.js
@@ -11,6 +11,14 @@ LitElement,
 html,
 css
 } from 'lit';
+import { renderButtonEditor } from '../cards/button/editor.js';
+import { renderPopUpEditor } from '../cards/pop-up/editor.js';
+import { renderSeparatorEditor } from '../cards/separator/editor.js';
+import { renderHorButtonStackEditor } from '../cards/horizontal-buttons-stack/editor.js';
+import { renderCoverEditor } from '../cards/cover/editor.js';
+import { renderClimateEditor } from '../cards/climate/editor.js';
+import { renderSelectEditor } from '../cards/select/editor.js';
+import { renderMediaPlayerEditor } from '../cards/media-player/editor.js';
 
 class BubbleCardEditor extends LitElement {
 
@@ -38,242 +46,6 @@ class BubbleCardEditor extends LitElement {
 
     get _entity() {
         return this._config?.entity || '';
-    }
-
-    get _name() {
-        return this._config?.name || '';
-    }
-
-    get _icon() {
-        return this._config?.icon || '';
-    }
-
-    get _state() {
-        return this._config?.state || '';
-    }
-
-    get _text() {
-        return this._config?.text || '';
-    }
-
-    get _hash() {
-        return this._config?.hash || '#pop-up-name';
-    }
-
-    get _trigger_entity() {
-        return this._config?.trigger_entity || '';
-    }
-
-    get _trigger_state() {
-        return this._config?.trigger_state || '';
-    }
-
-    get _trigger_close() {
-        return this._config?.trigger_close || false;
-    }
-
-    get _margin() {
-        return this._config?.margin || '7px';
-    }
-
-    get _margin_top_mobile() {
-        return this._config?.margin_top_mobile || '0px';
-    }
-
-    get _margin_top_desktop() {
-        return this._config?.margin_top_desktop || '0px';
-    }
-
-    get _width_desktop() {
-        return this._config?.width_desktop || '540px';
-    }
-
-    get _bg_color() {
-        return this._config?.bg_color || '';
-    }
-
-    get _bg_opacity() {
-        return this._config?.bg_opacity !== undefined ? this._config?.bg_opacity : '88';
-    }
-
-    get _bg_blur() {
-        return this._config?.bg_blur !== undefined ? this._config?.bg_blur : '10';
-    }
-
-    get _backdrop_blur() {
-        return this._config?.backdrop_blur !== undefined ? this._config?.backdrop_blur : '0';
-    }
-
-    get _shadow_opacity() {
-        return this._config?.shadow_opacity !== undefined ? this._config?.shadow_opacity : '0';
-    }
-
-    get _rise_animation() {
-        return this._config?.rise_animation !== undefined ? this._config?.rise_animation : true;
-    }
-
-    get _auto_close() {
-        return this._config?.auto_close || '';
-    }
-
-    get _close_on_click() {
-        return this._config?.close_on_click || false;
-    }
-
-    get _close_by_clicking_outside() {
-        return this._config?.close_by_clicking_outside ?? true;
-    }
-
-    get _background_update() {
-        return this._config?.background_update || false;
-    }
-
-    get _icon_open() {
-        return this._config?.icon_open || '';
-    }
-
-    get _icon_close() {
-        return this._config?.icon_close || '';
-    }
-
-    get _icon_down() {
-        return this._config?.icon_down || '';
-    }
-
-    get _icon_up() {
-        return this._config?.icon_up || '';
-    }
-
-    get _open_service() {
-        return this._config?.open_service || 'cover.open_cover';
-    }
-
-    get _close_service() {
-        return this._config?.close_service || 'cover.close_cover';
-    }
-
-    get _stop_service() {
-        return this._config?.stop_service || 'cover.stop_cover';
-    }
-
-    get _auto_order() {
-        return this._config?.auto_order || false;
-    }
-
-    get _highlight_current_view() {
-        return this._config?.highlight_current_view || false;
-    }
-
-    get _show_state() {
-        const defaultState = this._config?.card_type === 'state' ? true : false;
-        return this._config?.show_state || defaultState;
-    }
-
-    get _show_attribute() {
-        const defaultState = this._config.card_type === 'state' ? true : false;
-        return this._config.show_attribute || defaultState;
-    }
-
-    get _show_last_changed() {
-        const defaultState = this._config.card_type === 'state' ? true : false;
-        return this._config.show_last_changed || this._config.show_last_updated || defaultState;
-    }
-
-    get _attribute() {
-        return this._config.attribute || false;
-    }
-
-    get _hide_backdrop() {
-        return this._config.hide_backdrop ?? false;
-    }
-
-    get _hide_gradient() {
-        return this._config.hide_gradient || false;
-    }
-
-    get _hide_play_pause_button() {
-        return this._config.hide?.play_pause_button || false;
-    }
-
-    get _hide_next_button() {
-        return this._config.hide?.next_button || false;
-    }
-
-    get _hide_previous_button() {
-        return this._config.hide?.previous_button || false;
-    }
-
-    get _hide_volume_button() {
-        return this._config.hide?.volume_button || false;
-    }
-
-    get _hide_power_button() {
-        return this._config.hide?.power_button || false;
-    }  
-
-    get _sub_button() {
-        return this._config.sub_button || '';
-    }
-
-    get _button_action() {
-        return this._config.button_action || '';
-    }
-
-    get _open_action() {
-        return this._config.open_action || '';
-    }
-
-    get _close_action() {
-        return this._config.close_action || '';
-    }
-
-    get _show_header() {
-        return this._config.show_header ?? true;
-    }
-
-    get _slide_to_close_distance() {
-        return this._config.slide_to_close_distance ?? 400;
-    }
-
-    get _slider_live_update() {
-        return this._config.slider_live_update ?? false;
-    }
-
-    get _cover_background() {
-        return this._config.cover_background ?? false;
-    }
-
-    get _tap_action() {
-        return {
-            action: this._config.tap_action?.action || "more-info",
-            navigation_path: this._config.tap_action?.navigation_path || "",
-            url_path: this._config.tap_action?.url_path || "",
-            service: this._config.tap_action?.service || "",
-            target_entity: this._config.tap_action?.target?.entity_id || "",
-            data: this._config.tap_action?.data || ""
-        };
-    }
-
-    get _double_tap_action() {
-        return {
-            action: this._config.double_tap_action?.action || "toggle",
-            navigation_path: this._config.double_tap_action?.navigation_path || "",
-            url_path: this._config.double_tap_action?.url_path || "",
-            service: this._config.double_tap_action?.service || "",
-            target_entity: this._config.double_tap_action?.target?.entity_id || "",
-            data: this._config.double_tap_action?.data || ""
-        };
-    }
-
-    get _hold_action() {
-        return {
-            action: this._config.hold_action?.action || "toggle",
-            navigation_path: this._config.hold_action?.navigation_path || "",
-            url_path: this._config.hold_action?.url_path || "",
-            service: this._config.hold_action?.service || "",
-            target_entity: this._config.hold_action?.target?.entity_id || "",
-            data: this._config.hold_action?.data || ""
-        };
     }
 
     get _selectable_attributes() {
@@ -316,32 +88,6 @@ class BubbleCardEditor extends LitElement {
                 label: item,
                 value: item
             });
-
-            this.allEntitiesList = Object.keys(this.hass.states).map(formateList);
-
-            this.lightList = Object.keys(this.hass.states).filter(
-                (eid) => eid.substr(0, eid.indexOf(".")) === "light"
-            ).map(formateList);
-
-            this.sensorList = Object.keys(this.hass.states).filter(
-                (eid) => eid.substr(0, eid.indexOf(".")) === "sensor"
-            ).map(formateList);
-
-            this.binarySensorList = Object.keys(this.hass.states).filter(
-                (eid) => eid.substr(0, eid.indexOf(".")) === "binary_sensor"
-            ).map(formateList);
-
-            this.coverList = Object.keys(this.hass.states).filter(
-                (eid) => eid.substr(0, eid.indexOf(".")) === "cover"
-            ).map(formateList);
-
-            this.mediaPlayerList = Object.keys(this.hass.states).filter(
-                (eid) => eid.substr(0, eid.indexOf(".")) === "media_player"
-            ).map(formateList);
-
-            this.climateList = Object.keys(this.hass.states).filter(
-                (eid) => eid.substr(0, eid.indexOf(".")) === "climate"
-            ).map(formateList);
 
             this.inputSelectList = Object.keys(this.hass.states)
                 .filter((eid) => {
@@ -407,24 +153,6 @@ class BubbleCardEditor extends LitElement {
                 }
             ];
 
-            this.buttonTypeList = [{
-                    'label': 'Switch',
-                    'value': 'switch'
-                },
-                {
-                    'label': 'Slider',
-                    'value': 'slider'
-                },
-                {
-                    'label': 'State',
-                    'value': 'state'
-                },
-                {
-                    'label': 'Name / Text',
-                    'value': 'name'
-                }
-            ];
-
             this.tapActionTypeList = [{
                     'label': 'More info',
                     'value': 'more-info'
@@ -458,730 +186,21 @@ class BubbleCardEditor extends LitElement {
             this.listsUpdated = true;
         }
 
-        const allEntitiesList = this.allEntitiesList;
-        const lightList = this.lightList;
-        const sensorList = this.sensorList;
-        const coverList = this.coverList;
         const cardTypeList = this.cardTypeList;
         const buttonTypeList = this.buttonTypeList;
-        const nameButton = this._config?.button_type === 'name';
 
         if (this._config?.card_type === 'pop-up') {
-            const conditions = this._config?.trigger ?? [];
-            return html`
-                <div class="card-config">
-                    ${this.makeDropdown("Card type", "card_type", cardTypeList)}
-                    <ha-textfield
-                        label="Hash (e.g. #kitchen)"
-                        .value="${this._hash}"
-                        .configValue="${"hash"}"
-                        @input="${this._valueChanged}"
-                    ></ha-textfield>
-                    <ha-expansion-panel outlined>
-                        <h4 slot="header">
-                          <ha-icon icon="mdi:dock-top"></ha-icon>
-                          Header settings
-                        </h4>
-                        <div class="content">
-                            <ha-formfield .label="Optional - Show header">
-                                <ha-switch
-                                    aria-label="Optional - Show header"
-                                    .checked=${this._show_header}
-                                    .configValue="${"show_header"}"
-                                    @change=${this._valueChanged}
-                                ></ha-switch>
-                                <div class="mdc-form-field">
-                                    <label class="mdc-label">Optional - Show header</label> 
-                                </div>
-                            </ha-formfield>
-                            <ha-alert alert-type="info">You can completely hide the pop-up header, including the close button. To close it when hidden, either make a long swipe within the pop-up or click outside of it.</ha-alert>
-                            <div style="${!this._show_header ? 'display: none;' : ''}">
-                                <hr />
-                                ${this.makeDropdown("Button type", "button_type", buttonTypeList)}
-                                <ha-form
-                                    .hass=${this.hass}
-                                    .data=${this._config}
-                                    .schema=${[
-                                                { name: "entity",
-                                                label: "Optional - Entity", 
-                                                selector: { entity: {} },
-                                                },
-                                            ]}   
-                                    .computeLabel=${this._computeLabelCallback}
-                                    .disabled="${nameButton}"
-                                    @value-changed=${this._valueChanged}
-                                ></ha-form>                                         
-                                <ha-textfield
-                                    label="Optional - Name"
-                                    .value="${this._name}"
-                                    .configValue="${"name"}"
-                                    @input="${this._valueChanged}"
-                                ></ha-textfield>
-                                ${this.makeDropdown("Optional - Icon", "icon")}
-                                ${this.makeShowState()}
-                                <hr />
-                                <ha-expansion-panel outlined>
-                                    <h4 slot="header">
-                                      <ha-icon icon="mdi:gesture-tap"></ha-icon>
-                                      Tap action on icon
-                                    </h4>
-                                    <div class="content">
-                                        ${this.makeActionPanel("Tap action")}
-                                        ${this.makeActionPanel("Double tap action")}
-                                        ${this.makeActionPanel("Hold action")}
-                                    </div>
-                                </ha-expansion-panel>
-                                <ha-expansion-panel outlined style="display: ${this._config.button_type === 'slider' ? 'none' : ''}">
-                                    <h4 slot="header">
-                                      <ha-icon icon="mdi:gesture-tap"></ha-icon>
-                                      Tap action on button
-                                    </h4>
-                                    <div class="content">
-                                        ${this.makeActionPanel("Tap action", this._button_action, this._config.button_type !== 'name' ? (this._config.button_type === 'state' ? 'more-info' : 'toggle') : 'none', 'button_action')}
-                                        ${this.makeActionPanel("Double tap action", this._button_action, this._config.button_type !== 'name' ? (this._config.button_type === 'state' ? 'more-info' : 'toggle') : 'none', 'button_action')}
-                                        ${this.makeActionPanel("Hold action", this._button_action, this._config.button_type !== 'name' ? 'more-info' : 'none', 'button_action')}
-                                    </div>
-                                </ha-expansion-panel>
-                                ${this.makeSubButtonPanel()}
-                            </div>
-                        </div>
-                    </ha-expansion-panel>
-                    <ha-expansion-panel outlined>
-                        <h4 slot="header">
-                          <ha-icon icon="mdi:cog"></ha-icon>
-                          Pop-up settings
-                        </h4>
-                        <div class="content">
-                            <ha-textfield
-                                label="Optional - Auto close in milliseconds (e.g. 15000)"
-                                type="number"
-                                inputMode="numeric"
-                                min="0"
-                                step="1000"
-                                .value="${this._auto_close}"
-                                .configValue="${"auto_close"}"
-                                @input="${this._valueChanged}"
-                            ></ha-textfield>
-                            <ha-textfield
-                                label="Optional - Slide to close distance (default to 400)"
-                                type="number"
-                                inputMode="numeric"
-                                min="0"
-                                step="10"
-                                .value="${this._slide_to_close_distance}"
-                                .configValue="${"slide_to_close_distance"}"
-                                @input="${this._valueChanged}"
-                            ></ha-textfield>
-                            <ha-formfield .label="Optional - Close the pop-up by clicking outside of it (a refresh is needed)">
-                                <ha-switch
-                                    aria-label="Optional - Close the pop-up by clicking outside of it (a refresh is needed)"
-                                    .checked=${this._close_by_clicking_outside}
-                                    .configValue="${"close_by_clicking_outside"}"
-                                    @change=${this._valueChanged}
-                                ></ha-switch>
-                                <div class="mdc-form-field">
-                                    <label class="mdc-label">Optional - Close the pop-up by clicking outside of it (a refresh is needed)</label> 
-                                </div>
-                            </ha-formfield>
-                            <ha-formfield .label="Optional - Close the pop-up after any click or tap">
-                                <ha-switch
-                                    aria-label="Optional - Close the pop-up after any click or tap"
-                                    .checked=${this._close_on_click}
-                                    .configValue="${"close_on_click"}"
-                                    @change=${this._valueChanged}
-                                ></ha-switch>
-                                <div class="mdc-form-field">
-                                    <label class="mdc-label">Optional - Close the pop-up after any click or tap</label> 
-                                </div>
-                            </ha-formfield>
-                            <ha-formfield .label="Optional - Update cards in background (not recommended)">
-                                <ha-switch
-                                    aria-label="Optional - Update cards in background (not recommended)"
-                                    .checked=${this._background_update}
-                                    .configValue="${"background_update"}"
-                                    @change=${this._valueChanged}
-                                ></ha-switch>
-                                <div class="mdc-form-field">
-                                    <label class="mdc-label">Optional - Update cards in background (not recommended)</label> 
-                                </div>
-                            </ha-formfield>
-                            <ha-alert alert-type="info">Background updates are only recommended if you encounter issues with certain cards within your pop-up.</ha-alert>
-                        </div>
-                    </ha-expansion-panel>
-                    <ha-expansion-panel outlined>
-                        <h4 slot="header">
-                          <ha-icon icon="mdi:bell"></ha-icon>
-                          Pop-up trigger
-                        </h4>
-                        <div class="content">
-                            <ha-card-conditions-editor
-                                .hass=${this.hass}
-                                .conditions=${conditions}
-                                @value-changed=${(ev) => this._conditionChanged(ev)}
-                            >
-                            </ha-card-conditions-editor>
-                            <ha-alert alert-type="info">
-                                The pop-up will be opened when ALL conditions are fulfilled. For example you can open a "Security" pop-up with a camera when a person is in front of your house. You can also create a toggle helper (<code>input_boolean</code>) and trigger its opening/closing in an automation.
-                            </ha-alert>
-                        </div>
-                    </ha-expansion-panel>
-                    <ha-expansion-panel outlined>
-                        <h4 slot="header">
-                          <ha-icon icon="mdi:gesture-tap"></ha-icon>
-                          Pop-up open/close action
-                        </h4>
-                        <div class="content">
-                            ${this.makeActionPanel("Open action", this._config, 'none')}
-                            ${this.makeActionPanel("Close action", this._config, 'none')}
-                            <ha-alert alert-type="info">This allows you to trigger an action on pop-up open/close.</ha-alert>
-                        </div>
-                    </ha-expansion-panel>
-                    <ha-expansion-panel outlined>
-                        <h4 slot="header">
-                          <ha-icon icon="mdi:palette"></ha-icon>
-                          Styling options
-                        </h4>
-                        <div class="content">
-                            ${this.makeLayoutOptions()}
-                            <ha-expansion-panel outlined>
-                                <h4 slot="header">
-                                  <ha-icon icon="mdi:palette"></ha-icon>
-                                  Pop-up styling
-                                </h4>
-                                <div class="content"> 
-                                    <ha-textfield
-                                        label="Optional - Margin (fix centering on some themes) (e.g. 13px)"
-                                        .value="${this._margin}"
-                                        .configValue="${"margin"}"
-                                        @input="${this._valueChanged}"
-                                    ></ha-textfield>
-                                    <ha-textfield
-                                        label="Optional - Top margin on mobile (e.g. -56px if your header is hidden)"
-                                        .value="${this._margin_top_mobile}"
-                                        .configValue="${"margin_top_mobile"}"
-                                        @input="${this._valueChanged}"
-                                    ></ha-textfield>
-                                    <ha-textfield
-                                        label="Optional - Top margin on desktop (e.g. 50vh for an half sized pop-up)"
-                                        .value="${this._margin_top_desktop}"
-                                        .configValue="${"margin_top_desktop"}"
-                                        @input="${this._valueChanged}"
-                                    ></ha-textfield>
-                                    <ha-textfield
-                                        label="Optional - Width on desktop (100% by default on mobile)"
-                                        .value="${this._width_desktop}"
-                                        .configValue="${"width_desktop"}"
-                                        @input="${this._valueChanged}"
-                                    ></ha-textfield>
-                                    <ha-textfield
-                                        label="Optional - Background color (any var, hex, rgb or rgba value)"
-                                        .value="${this._bg_color}"
-                                        .configValue="${"bg_color"}"
-                                        @input="${this._valueChanged}"
-                                    ></ha-textfield>
-                                    <ha-textfield
-                                        label="Optional - Background opacity (0-100 range)"
-                                        type="number"
-                                        inputMode="numeric"
-                                        min="0"
-                                        max="100"
-                                        .value="${this._bg_opacity}"
-                                        .configValue="${"bg_opacity"}"
-                                        @input="${this._valueChanged}"
-                                    ></ha-textfield>
-                                    <ha-textfield
-                                        label="Optional - Background blur (0-100 range)"
-                                        type="number"
-                                        inputMode="numeric"
-                                        min="0"
-                                        max="100"
-                                        .value="${this._bg_blur}"
-                                        .configValue="${"bg_blur"}"
-                                        @input="${this._valueChanged}"
-                                    ></ha-textfield>
-                                    <ha-textfield
-                                        label="Optional - Backdrop blur (0-100 range)"
-                                        type="number"
-                                        inputMode="numeric"
-                                        min="0"
-                                        max="100"
-                                        .value="${this._backdrop_blur}"
-                                        .configValue="${"backdrop_blur"}"
-                                        @input="${this._valueChanged}"
-                                    ></ha-textfield>
-                                    <ha-textfield
-                                        label="Optional - Shadow opacity (0-100 range)"
-                                        type="number"
-                                        inputMode="numeric"
-                                        min="0"
-                                        max="100"
-                                        .configValue="${"shadow_opacity"}"
-                                        .value="${this._shadow_opacity}""
-                                        @input="${this._valueChanged}"
-                                    ></ha-textfield>
-                                    <ha-formfield .label="Optional - Hide pop-up backdrop (a refresh is needed)">
-                                        <ha-switch
-                                            aria-label="Optional - Hide pop-up backdrop (a refresh is needed)"
-                                            .checked=${this._hide_backdrop}
-                                            .configValue="${"hide_backdrop"}"
-                                            @change=${this._valueChanged}
-                                        ></ha-switch>
-                                        <div class="mdc-form-field">
-                                            <label class="mdc-label">Optional - Hide pop-up backdrop (a refresh is needed)</label> 
-                                        </div>
-                                    </ha-formfield>
-                                    <ha-alert alert-type="warning">Set this toggle to true on the first pop-up of your main dashboard to hide the darker backdrop behind all pop-ups. <b>You can add a blurred effect to it by changing <code>Optional - Backdrop blur</code> just below, but be aware that this can slow down your dashboard when opening pop-ups. It is now set to 0 for that reason.</b></ha-alert>
-                                </div>
-                            </ha-expansion-panel>
-                            ${this.makeStyleEditor()}
-                        </div>
-                    </ha-expansion-panel>
-                    <ha-alert alert-type="info">
-                        This card allows you to convert any vertical stack into a pop-up. Each pop-up is hidden by default and can be opened by targeting its link (e.g., '#pop-up-name'), with <a style="color: var(--text-primary-color)" href="https://github.com/Clooos/Bubble-Card#example">any card</a> that supports the <code>navigate</code> action, or with the <a style="color: var(--text-primary-color)" href="https://github.com/Clooos/Bubble-Card#horizontal-buttons-stack">horizontal buttons stack</a> that is included.
-                        <br><br><b>Important:</b> This card must be placed within a <a style="color: var(--text-primary-color)" href="https://www.home-assistant.io/dashboards/vertical-stack/">vertical stack</a> card at the topmost position to function properly. To avoid misalignment with your view, place vertical stacks/pop-ups after all other dashboard cards. It should be called from the same view to work. 
-                        <br><br><b>You can also watch this <a style="color: var(--text-primary-color)" href="https://www.youtube.com/watch?v=7mOV7BfWoFc">video</a> that explains how to create your first pop-up.</b>
-                    </ha-alert>
-                    <ha-alert alert-type="warning">Since v1.7.0, the optimized mode has been removed to ensure stability and to simplify updates for everyone. However, if your pop-up content still appears on the screen during page loading, <a style="color: var(--text-primary-color)" href="https://github.com/Clooos/Bubble-Card#pop-up-initialization-fix">you can install this similar fix.</a></ha-alert>
-                    ${this.makeVersion()}
-              </div>
-            `;
+            return renderPopUpEditor(this);
         } else if (this._config?.card_type === 'button') {
-            let entityList = {}
-            if (this._button_type === 'slider') entityList = {domain:["light", "media_player", "cover","input_number"]};
-
-            return html`
-                <div class="card-config">
-                    ${this.makeDropdown("Card type", "card_type", cardTypeList)}
-                    ${this.makeDropdown("Button type", "button_type", buttonTypeList)}
-                    <ha-form
-                        .hass=${this.hass}
-                        .data=${this._config}
-                        .schema=${[
-                                    { name: "entity",
-                                    label: this._button_type !== 'slider' ? "Entity (toggle)" : "Entity (light, media_player, cover or input_number)", 
-                                    selector: { entity: entityList },
-                                    },
-                                ]}   
-                        .computeLabel=${this._computeLabelCallback}
-                        .disabled="${nameButton}"
-                        @value-changed=${this._valueChanged}
-                    ></ha-form>
-                    <ha-expansion-panel outlined>
-                        <h4 slot="header">
-                          <ha-icon icon="mdi:cog"></ha-icon>
-                          Button settings
-                        </h4>
-                        <div class="content">     
-                            <ha-textfield
-                                label="Optional - Name"
-                                .value="${this._name}"
-                                .configValue="${"name"}"
-                                @input="${this._valueChanged}"
-                            ></ha-textfield>
-                            ${this.makeDropdown("Optional - Icon", "icon")}
-                            ${this.makeShowState()}
-                        </div>
-                    </ha-expansion-panel>
-                    <ha-expansion-panel outlined>
-                        <h4 slot="header">
-                          <ha-icon icon="mdi:gesture-tap"></ha-icon>
-                          Tap action on icon
-                        </h4>
-                        <div class="content">
-                            ${this.makeActionPanel("Tap action")}
-                            ${this.makeActionPanel("Double tap action")}
-                            ${this.makeActionPanel("Hold action")}
-                        </div>
-                    </ha-expansion-panel>
-                    <ha-expansion-panel outlined style="display: ${this._config.button_type === 'slider' ? 'none' : ''}">
-                        <h4 slot="header">
-                          <ha-icon icon="mdi:gesture-tap"></ha-icon>
-                          Tap action on button
-                        </h4>
-                        <div class="content">
-                            ${this.makeActionPanel("Tap action", this._button_action, this._config.button_type !== 'name' ? (this._config.button_type === 'state' ? 'more-info' : 'toggle') : 'none', 'button_action')}
-                            ${this.makeActionPanel("Double tap action", this._button_action, this._config.button_type !== 'name' ? (this._config.button_type === 'state' ? 'more-info' : 'toggle') : 'none', 'button_action')}
-                            ${this.makeActionPanel("Hold action", this._button_action, this._config.button_type !== 'name' ? 'more-info' : 'none', 'button_action')}
-                        </div>
-                    </ha-expansion-panel>
-                    <ha-expansion-panel outlined>
-                        <h4 slot="header">
-                          <ha-icon icon="mdi:palette"></ha-icon>
-                          Styling options
-                        </h4>
-                        <div class="content">
-                            ${this.makeLayoutOptions()}
-                            ${this.makeStyleEditor()}
-                        </div>
-                    </ha-expansion-panel>
-                    ${this.makeSubButtonPanel()}
-                    <ha-alert alert-type="info">This card allows you to control your entities. ${this._config.button_type === 'slider' ? 'Supported entities: Light (brightness), media player (volume), cover (position), fan (percentage), climate (temperature), input number and number (value). To access color / control of an entity, simply tap on the icon.' : ''}</ha-alert>
-                    ${this.makeVersion()}
-                </div>
-            `;
+            return renderButtonEditor(this);
         } else if (this._config?.card_type === 'separator') {
-            return html`
-                <div class="card-config">
-                    ${this.makeDropdown("Card type", "card_type", cardTypeList)}
-                    <ha-textfield
-                        label="Name"
-                        .value="${this._name}"
-                        .configValue="${"name"}"
-                        @input="${this._valueChanged}"
-                    ></ha-textfield>
-                    ${this.makeDropdown("Icon", "icon")}
-                    <ha-expansion-panel outlined>
-                        <h4 slot="header">
-                          <ha-icon icon="mdi:palette"></ha-icon>
-                          Styling options
-                        </h4>
-                        <div class="content">
-                            ${this.makeLayoutOptions()}
-                            ${this.makeStyleEditor()}
-                        </div>
-                    </ha-expansion-panel>
-                    ${this.makeSubButtonPanel()}
-                    <ha-alert alert-type="info">This card is a simple separator for dividing your pop-up into categories / sections. e.g. Lights, Devices, Covers, Settings, Automations...</ha-alert>
-                    ${this.makeVersion()}
-              </div>
-            `;
+            return renderSeparatorEditor(this);
         } else if (this._config?.card_type === 'horizontal-buttons-stack') {
-            if (!this.buttonAdded) {
-                this.buttonAdded = true;
-                this.buttonIndex = 0;
-
-                while (this._config[(this.buttonIndex + 1) + '_link']) {
-                    this.buttonIndex++;
-                }
-            }
-
-            function addButton() {
-                this.buttonIndex++;
-                this.requestUpdate();
-            }
-
-            return html`
-                <div class="card-config">
-                    ${this.makeDropdown("Card type", "card_type", cardTypeList)}
-                    <div id="buttons-container">
-                        ${this.makeButton()}
-                    </div>
-                    <button class="icon-button" @click="${addButton}">
-                        <ha-icon icon="mdi:plus"></ha-icon>
-                        New button
-                    </button>
-                    <ha-formfield .label="Auto order">
-                        <ha-switch
-                            aria-label="Toggle auto order"
-                            .checked=${this._auto_order}
-                            .configValue="${"auto_order"}"
-                            @change=${this._valueChanged}
-                        ></ha-switch>
-                        <div class="mdc-form-field">
-                            <label class="mdc-label">Optional - Auto order (Presence/occupancy sensors needed)</label> 
-                        </div>
-                    </ha-formfield>
-                    <ha-expansion-panel outlined>
-                        <h4 slot="header">
-                          <ha-icon icon="mdi:palette"></ha-icon>
-                          Styling options
-                        </h4>
-                        <div class="content">  
-                            <ha-expansion-panel outlined>
-                                <h4 slot="header">
-                                  <ha-icon icon="mdi:palette"></ha-icon>
-                                  Horizontal buttons stack styling
-                                </h4>
-                                <div class="content"> 
-                                    <ha-textfield
-                                        label="Optional - Margin (fix centering on some themes) (e.g. 13px)"
-                                        .value="${this._margin}"
-                                        .configValue="${"margin"}"
-                                        @input="${this._valueChanged}"
-                                    ></ha-textfield>
-                                    <ha-textfield
-                                        label="Optional - Width on desktop (100% by default on mobile)"
-                                        .value="${this._width_desktop}"
-                                        .configValue="${"width_desktop"}"
-                                        @input="${this._valueChanged}"
-                                    ></ha-textfield>
-                                    <ha-formfield .label="Optional - Rise animation (Displays an animation once the page has loaded)">
-                                        <ha-switch
-                                            aria-label="Optional - Rise animation (Displays an animation once the page has loaded)"
-                                            .checked=${this._rise_animation}
-                                            .configValue="${"rise_animation"}"
-                                            @change=${this._valueChanged}
-                                        ></ha-switch>
-                                        <div class="mdc-form-field">
-                                            <label class="mdc-label">Optional - Rise animation (Displays an animation once the page has loaded)</label> 
-                                        </div>
-                                    </ha-formfield>
-                                    <ha-formfield .label="Optional - Highlight current hash / view">
-                                        <ha-switch
-                                            aria-label="Optional - Highlight current hash / view"
-                                            .checked=${this._highlight_current_view}
-                                            .configValue="${"highlight_current_view"}"
-                                            @change=${this._valueChanged}
-                                        ></ha-switch>
-                                        <div class="mdc-form-field">
-                                            <label class="mdc-label">Optional - Highlight current hash / view</label> 
-                                        </div>
-                                    </ha-formfield>
-                                    <ha-formfield .label="Optional - Hide gradient">
-                                        <ha-switch
-                                            aria-label="Optional - Hide gradient"
-                                            .checked=${this._hide_gradient}
-                                            .configValue="${"hide_gradient"}"
-                                            @change=${this._valueChanged}
-                                        ></ha-switch>
-                                        <div class="mdc-form-field">
-                                            <label class="mdc-label">Optional - Hide gradient</label> 
-                                        </div>
-                                    </ha-formfield>
-                                </div>
-                            </ha-expansion-panel>
-                            ${this.makeStyleEditor()}
-                        </div>
-                    </ha-expansion-panel>
-                    <ha-alert alert-type="info">This card is the companion to the pop-up card, allowing you to open the corresponding pop-ups. It also allows you to open any page of your dashboard. In addition, you can add your motion sensors so that the order of the buttons adapts according to the room you just entered. This card is scrollable, remains visible and acts as a footer.</ha-alert>
-                    ${this.makeVersion()}
-                </div>
-            `;
+            return renderHorButtonStackEditor(this);
         } else if (this._config?.card_type === 'cover') {
-            return html`
-                <div class="card-config">
-                    ${this.makeDropdown("Card type", "card_type", cardTypeList)}
-                    <ha-form
-                        .hass=${this.hass}
-                        .data=${this._config}
-                        .schema=${[
-                                    { name: "entity",
-                                    label: "Entity", 
-                                    selector: { entity: {domain:["cover"]}  },
-                                    },
-                                ]}   
-                        .computeLabel=${this._computeLabelCallback}
-                        @value-changed=${this._valueChanged}
-                    ></ha-form>
-                    <ha-expansion-panel outlined>
-                        <h4 slot="header">
-                          <ha-icon icon="mdi:cog"></ha-icon>
-                          Cover settings
-                        </h4>
-                        <div class="content"> 
-                            <ha-textfield
-                                label="Optional - Name"
-                                .value="${this._name || ''}"
-                                .configValue="${"name"}"
-                                @input="${this._valueChanged}"
-                            ></ha-textfield>
-                            ${this.makeDropdown("Optional - Open icon", "icon_open")}
-                            ${this.makeDropdown("Optional - Closed icon", "icon_close")}
-                            ${this.makeShowState()}
-                        </div>
-                    </ha-expansion-panel>
-                    <ha-expansion-panel outlined>
-                        <h4 slot="header">
-                          <ha-icon icon="mdi:window-shutter-cog"></ha-icon>
-                          Custom services
-                        </h4>
-                        <div class="content"> 
-                            <ha-textfield
-                                label="Optional - Open service (cover.open_cover by default)"
-                                .value="${this._open_service}"
-                                .configValue="${"open_service"}"
-                                @input="${this._valueChanged}"
-                            ></ha-textfield>
-                            <ha-textfield
-                                label="Optional - Stop service (cover.stop_cover by default)"
-                                .value="${this._stop_service}"
-                                .configValue="${"stop_service"}"
-                                @input="${this._valueChanged}"
-                            ></ha-textfield>
-                            <ha-textfield
-                                label="Optional - Close service (cover.close_cover by default)"
-                                .value="${this._close_service}"
-                                .configValue="${"close_service"}"
-                                @input="${this._valueChanged}"
-                            ></ha-textfield>
-                        </div>
-                    </ha-expansion-panel>
-                    <ha-expansion-panel outlined>
-                        <h4 slot="header">
-                          <ha-icon icon="mdi:gesture-tap"></ha-icon>
-                          Tap action on icon
-                        </h4>
-                        <div class="content">
-                            ${this.makeActionPanel("Tap action")}
-                            ${this.makeActionPanel("Double tap action")}
-                            ${this.makeActionPanel("Hold action")}
-                        </div>
-                    </ha-expansion-panel>
-                    <ha-expansion-panel outlined>
-                        <h4 slot="header">
-                          <ha-icon icon="mdi:palette"></ha-icon>
-                          Styling options
-                        </h4>
-                        <div class="content"> 
-                            ${this.makeLayoutOptions()}
-                            <ha-expansion-panel outlined>
-                                <h4 slot="header">
-                                  <ha-icon icon="mdi:palette"></ha-icon>
-                                  Cover styling
-                                </h4>
-                                <div class="content"> 
-                                    ${this.makeDropdown("Optional - Arrow down icon", "icon_down")}
-                                    ${this.makeDropdown("Optional - Arrow up icon", "icon_up")}
-                                </div>
-                            </ha-expansion-panel>
-                            ${this.makeStyleEditor()}
-                        </div>
-                    </ha-expansion-panel>
-                    ${this.makeSubButtonPanel()}
-                    <ha-alert alert-type="info">This card allows you to control your covers.</ha-alert>
-                    ${this.makeVersion()}
-                </div>
-            `;
+            return renderCoverEditor(this);
         } else if (this._config?.card_type === 'media-player') {
-            return html`
-                <div class="card-config">
-                    ${this.makeDropdown("Card type", "card_type", cardTypeList)}
-                    <ha-form
-                        .hass=${this.hass}
-                        .data=${this._config}
-                        .schema=${[
-                                    { name: "entity",
-                                    label: "Entity", 
-                                    selector: { entity: {domain:["media_player"]}  },
-                                    },
-                                ]}   
-                        .computeLabel=${this._computeLabelCallback}
-                        @value-changed=${this._valueChanged}
-                    ></ha-form>
-                    <ha-expansion-panel outlined>
-                        <h4 slot="header">
-                          <ha-icon icon="mdi:cog"></ha-icon>
-                          Media player settings
-                        </h4>
-                        <div class="content"> 
-                            <ha-textfield
-                                label="Optional - Name"
-                                .value="${this._name || ''}"
-                                .configValue="${"name"}"
-                                @input="${this._valueChanged}"
-                            ></ha-textfield>
-                            ${this.makeDropdown("Optional - Icon", "icon")}
-                            ${this.makeShowState()}
-                        </div>
-                    </ha-expansion-panel>
-                    <ha-expansion-panel outlined>
-                        <h4 slot="header">
-                          <ha-icon icon="mdi:eye-off"></ha-icon>
-                          Display/hide buttons
-                        </h4>
-                        <div class="content"> 
-                            <ha-formfield .label="Optional - Hide play/pause button">
-                                <ha-switch
-                                    aria-label="Optional - Hide play/pause button"
-                                    .checked=${this._hide_play_pause_button}
-                                    .configValue="${"hide.play_pause_button"}"
-                                    @change=${this._valueChanged}
-                                ></ha-switch>
-                                <div class="mdc-form-field">
-                                    <label class="mdc-label">Optional - Hide play/pause button</label> 
-                                </div>
-                            </ha-formfield>
-                            <ha-formfield .label="Optional - Hide volume button">
-                                <ha-switch
-                                    aria-label="Optional - Hide volume button"
-                                    .checked=${this._hide_volume_button}
-                                    .configValue="${"hide.volume_button"}"
-                                    @change=${this._valueChanged}
-                                ></ha-switch>
-                                <div class="mdc-form-field">
-                                    <label class="mdc-label">Optional - Hide volume button</label>
-                                </div>
-                            </ha-formfield>
-                            <ha-formfield .label="Optional - Hide next button">
-                                <ha-switch
-                                    aria-label="Optional - Hide next button"
-                                    .checked=${this._hide_next_button}
-                                    .configValue="${"hide.next_button"}"
-                                    @change=${this._valueChanged}
-                                ></ha-switch>
-                                <div class="mdc-form-field">
-                                    <label class="mdc-label">Optional - Hide next button</label>
-                                </div>
-                            </ha-formfield>
-                            <ha-formfield .label="Optional - Hide previous button">
-                                <ha-switch
-                                    aria-label="Optional - Hide previous button"
-                                    .checked=${this._hide_previous_button}
-                                    .configValue="${"hide.previous_button"}"
-                                    @change=${this._valueChanged}
-                                ></ha-switch>
-                                <div class="mdc-form-field">
-                                    <label class="mdc-label">Optional - Hide previous button</label>
-                                </div>
-                            </ha-formfield>
-                            <ha-formfield .label="Optional - Hide power button">
-                                <ha-switch
-                                    aria-label="Optional - Hide power button"
-                                    .checked=${this._hide_power_button}
-                                    .configValue="${"hide.power_button"}"
-                                    @change=${this._valueChanged}
-                                ></ha-switch>
-                                <div class="mdc-form-field">
-                                    <label class="mdc-label">Optional - Hide power button</label>
-                                </div>
-                            </ha-formfield>
-                        </div>
-                    </ha-expansion-panel>
-                    <ha-expansion-panel outlined>
-                        <h4 slot="header">
-                          <ha-icon icon="mdi:gesture-tap"></ha-icon>
-                          Tap action on icon
-                        </h4>
-                        <div class="content">
-                            ${this.makeActionPanel("Tap action")}
-                            ${this.makeActionPanel("Double tap action")}
-                            ${this.makeActionPanel("Hold action")}
-                        </div>
-                    </ha-expansion-panel>
-                    <ha-expansion-panel outlined>
-                        <h4 slot="header">
-                          <ha-icon icon="mdi:palette"></ha-icon>
-                          Styling options
-                        </h4>
-                        <div class="content">
-                            ${this.makeLayoutOptions()}
-                            <ha-expansion-panel outlined>
-                                <h4 slot="header">
-                                  <ha-icon icon="mdi:palette"></ha-icon>
-                                  Media player styling
-                                </h4>
-                                <div class="content"> 
-                                    <ha-formfield .label="Optional - Blurred media cover in background">
-                                        <ha-switch
-                                            aria-label="Optional - Blurred media cover in background"
-                                            .checked=${this._cover_background}
-                                            .configValue="${"cover_background"}"
-                                            @change=${this._valueChanged}
-                                        ></ha-switch>
-                                        <div class="mdc-form-field">
-                                            <label class="mdc-label">Optional - Blurred media cover in background</label> 
-                                        </div>
-                                    </ha-formfield>
-                                </div>
-                            </ha-expansion-panel>
-                            ${this.makeStyleEditor()}
-                        </div>
-                    </ha-expansion-panel>
-                    ${this.makeSubButtonPanel()}
-                    <ha-alert alert-type="info">This card allows you to control a media player. You can tap on the icon to get more control.</ha-alert>
-                    ${this.makeVersion()}
-                </div>
-            `;
+            return renderMediaPlayerEditor(this);
         } else if (this._config?.card_type === 'empty-column') {
             return html`
                 <div class="card-config">
@@ -1200,208 +219,9 @@ class BubbleCardEditor extends LitElement {
                 </div>
             `;
         } else if (this._config?.card_type === 'select') {
-            const entity = this._config.entity;
-            const isSelect = entity?.startsWith("input_select") || entity?.startsWith("select") || this._config.select_attribute;
-            const entityAttribute = this.hass.states[entity]?.attributes;
-            const hasSelectAttributeList = this._selectable_attributes.some(attr => entityAttribute?.[attr]);
-            const selectableAttributeList = Object.keys(this.hass.states[entity]?.attributes || {}).map((attributeName) => {
-                let state = this.hass.states[entity];
-                let formattedName = this.hass.formatEntityAttributeName(state, attributeName);
-                return { label: formattedName, value: attributeName };
-            }).filter(attribute => this._selectable_attributes.includes(attribute.value));
-
-            return html`
-                <div class="card-config">
-                    ${this.makeDropdown("Card type", "card_type", cardTypeList)}
-                    <ha-form
-                        .hass=${this.inputSelectList2}
-                        .data=${this._config}
-                        .schema=${[
-                                    { name: "entity",
-                                    label: "Entity", 
-                                    selector: { entity: {}},
-                                    },
-                                ]}   
-                        .computeLabel=${this._computeLabelCallback}
-                        @value-changed=${this._valueChanged}
-                    ></ha-form>
-                    ${hasSelectAttributeList ? html`
-                        <div class="ha-combo-box">
-                            <ha-combo-box
-                                label="Select menu (from attributes)"
-                                .value="${this._config.select_attribute}"
-                                .items="${selectableAttributeList}"
-                                .configValue="${"select_attribute"}"
-                                @value-changed="${this._valueChanged}"
-                            ></ha-combo-box>
-                        </div>
-                    ` : ''}
-                    <ha-expansion-panel outlined>
-                        <h4 slot="header">
-                          <ha-icon icon="mdi:cog"></ha-icon>
-                          Button settings
-                        </h4>
-                        <div class="content">                   
-                            <ha-textfield
-                                label="Optional - Name"
-                                .value="${this._name}"
-                                .configValue="${"name"}"
-                                @input="${this._valueChanged}"
-                            ></ha-textfield>
-                            ${this.makeDropdown("Optional - Icon", "icon")}
-                            ${this.makeShowState()}
-                        </div>
-                    </ha-expansion-panel>
-                    <ha-expansion-panel outlined>
-                        <h4 slot="header">
-                          <ha-icon icon="mdi:gesture-tap"></ha-icon>
-                          Tap action on icon
-                        </h4>
-                        <div class="content">
-                            ${this.makeActionPanel("Tap action")}
-                            ${this.makeActionPanel("Double tap action")}
-                            ${this.makeActionPanel("Hold action")}
-                        </div>
-                    </ha-expansion-panel>
-                    <ha-expansion-panel outlined>
-                        <h4 slot="header">
-                          <ha-icon icon="mdi:palette"></ha-icon>
-                          Styling options
-                        </h4>
-                        <div class="content">
-                            ${this.makeLayoutOptions()}
-                            ${this.makeStyleEditor()}
-                        </div>
-                    </ha-expansion-panel>
-                    ${this.makeSubButtonPanel()}
-                    <ha-alert alert-type="info">
-                      This card allows you to have a select menu for your 
-                      <code>input_select</code>, <code>select</code> entities, and 
-                      any other entities that have attribute lists like 
-                      <code>source_list</code>, <code>sound_mode_list</code>, 
-                      <code>hvac_modes</code>, <code>fan_modes</code>, 
-                      <code>swing_modes</code>, <code>preset_modes</code>, or 
-                      <code>effect_list</code>.
-                    </ha-alert>
-                    ${this.makeVersion()}
-                </div>
-            `;
+            return renderSelectEditor(this);
         } else if (this._config?.card_type === 'climate') {
-            if (
-                this._config.card_type === "climate" && 
-                !this.climateSubButtonsAdded &&
-                this._config.entity
-            ) {
-                const shouldAddHVACModes = this.hass.states[this._config.entity]?.attributes?.hvac_modes;
-
-                if (!this._config.sub_button || this._config.sub_button.length === 0) {
-                    this._config.sub_button = [
-                        shouldAddHVACModes ? { 
-                            name: 'HVAC modes menu', 
-                            select_attribute: 'hvac_modes', 
-                            state_background: false,
-                            show_arrow: false 
-                        } : null
-                    ].filter(Boolean);
-                }
-
-                this.climateSubButtonsAdded = true;
-            }
-
-            return html`
-                <div class="card-config">
-                ${this.makeDropdown("Card type", "card_type", cardTypeList)}
-                <ha-form
-                    .hass=${this.hass}
-                    .data=${this._config}
-                    .schema=${[
-                                { name: "entity",
-                                label: "Entity", 
-                                selector: { entity: {domain:["climate"]}  },
-                                },
-                            ]}   
-                    .computeLabel=${this._computeLabelCallback}
-                    @value-changed=${this._valueChanged}
-                ></ha-form>
-                                        <ha-expansion-panel outlined>
-                        <h4 slot="header">
-                          <ha-icon icon="mdi:cog"></ha-icon>
-                          Climate settings
-                        </h4>
-                        <div class="content">     
-                            <ha-textfield
-                                label="Optional - Name"
-                                .value="${this._name}"
-                                .configValue="${"name"}"
-                                @input="${this._valueChanged}"
-                            ></ha-textfield>
-                            ${this.makeDropdown("Optional - Icon", "icon")}
-                            ${this.makeShowState()}
-                            ${this.hass.states[this._config.entity]?.attributes?.target_temp_low ? html`
-                                <ha-formfield .label="Optional - Hide target temp low">
-                                    <ha-switch
-                                        aria-label="Optional - Hide target temp low"
-                                        .checked=${this._config.hide_target_temp_low}
-                                        .configValue="${"hide_target_temp_low"}"
-                                        @change=${this._valueChanged}
-                                    ></ha-switch>
-                                    <div class="mdc-form-field">
-                                        <label class="mdc-label">Optional - Hide target temp low</label> 
-                                    </div>
-                                </ha-formfield>
-                            ` : ''}
-                            ${this.hass.states[this._config.entity]?.attributes?.target_temp_high ? html`
-                                <ha-formfield .label="Optional - Hide target temp high">
-                                    <ha-switch
-                                        aria-label="Optional - Hide target temp high"
-                                        .checked=${this._config.hide_target_temp_high}
-                                        .configValue="${"hide_target_temp_high"}"
-                                        @change=${this._valueChanged}
-                                    ></ha-switch>
-                                    <div class="mdc-form-field">
-                                        <label class="mdc-label">Optional - Hide target temp high</label> 
-                                    </div>
-                                </ha-formfield>
-                            ` : ''}
-                            <ha-formfield .label="Optional - Constant background color when ON">
-                                <ha-switch
-                                    aria-label="Optional - Constant background color when ON"
-                                    .checked=${this._config.state_color === true}
-                                    .configValue="${"state_color"}"
-                                    @change=${this._valueChanged}
-                                ></ha-switch>
-                                <div class="mdc-form-field">
-                                    <label class="mdc-label">Optional - Constant background color when ON</label> 
-                                </div>
-                            </ha-formfield>
-                        </div>
-                    </ha-expansion-panel>
-                    <ha-expansion-panel outlined>
-                        <h4 slot="header">
-                          <ha-icon icon="mdi:gesture-tap"></ha-icon>
-                          Tap action on icon
-                        </h4>
-                        <div class="content">
-                            ${this.makeActionPanel("Tap action")}
-                            ${this.makeActionPanel("Double tap action")}
-                            ${this.makeActionPanel("Hold action")}
-                        </div>
-                    </ha-expansion-panel>
-                    <ha-expansion-panel outlined>
-                        <h4 slot="header">
-                          <ha-icon icon="mdi:palette"></ha-icon>
-                          Styling options
-                        </h4>
-                        <div class="content">
-                            ${this.makeLayoutOptions()}
-                            ${this.makeStyleEditor()}
-                        </div>
-                    </ha-expansion-panel>
-                    ${this.makeSubButtonPanel()}
-                    <ha-alert alert-type="info">This card allows you to control your climate entities. You can also add a sub-button that display a select menu for your climate modes (check if you have "Select menu" available when you create a new sub-button).</ha-alert>
-                    ${this.makeVersion()}
-                </div>
-            `;
+            return renderClimateEditor(this);
         } else if (!this._config?.card_type) {
             return html`
                 <div class="card-config">
@@ -1642,7 +462,7 @@ class BubbleCardEditor extends LitElement {
                 <ha-formfield .label="Optional - Slider live update">
                     <ha-switch
                         aria-label="Optional - Slider live update"
-                        .checked=${this._slider_live_update}
+                        .checked=${this._config.slider_live_update ?? false}
                         .configValue="${"slider_live_update"}"
                         @change=${this._valueChanged}
                     ></ha-switch>
